% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_formatter.R
\name{number_formatter}
\alias{number_formatter}
\title{Number formatter function}
\usage{
number_formatter(number, rounding = FALSE, digits = ifelse(rounding,
  NA, 6), currency = FALSE)
}
\arguments{
\item{number}{A number you want to format - if it is a vector you'll have to map the function to the vector, e.g.
map(vector, number_formatter, rounding = TRUE, digits = 2, currency = TRUE).}

\item{rounding}{determines if number gets rounded. Defaults to FALSE.}

\item{digits}{determines the number of decimals for rounding.}

\item{currency}{determines if the $ sign is prefixed. Defaults to FALSE.}
}
\description{
This function allows you to add a negative "(-)" prefix as well as number suffixes, such as k for thousands, or M for millions.
}
