% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_formatter.R
\name{suffix_formatter_0}
\alias{suffix_formatter_0}
\title{Number formatter function}
\usage{
suffix_formatter_0(x, digits = NULL)
}
\arguments{
\item{x}{A vector of numbers representing an axis in your ggplot.}

\item{digits}{Defaults to NULL. Here it set to zero for use in plots.}
}
\description{
This function allows you to add number suffixes, such as k for thousands,
or M for millions to ggplots. It does NOT return anything when called by itself. Example scale_y_continuous(labels = suffix_formatter_0)
}
\details{
Based on this stackoverflow post https://stackoverflow.com/questions/28159936/formatting-large-currency-or-dollar-values-to-millions-billions
}
