% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_formatter_0_cur.R
\name{suffix_formatter_0_cur}
\alias{suffix_formatter_0_cur}
\title{Number formatter function}
\usage{
suffix_formatter_0_cur(x, digits = NULL)
}
\arguments{
\item{x}{A vector of numbers representing an axis in your ggplot.}

\item{digits}{Defaults to NULL. Here it set to zero for use in plots.}
}
\description{
Same as number_formatter but for currency $.
Example scale_y_continuous(labels = suffix_formatter_0_cur)
}
\details{
Rewritten for currency from Based on this stackoverflow post https://stackoverflow.com/questions/28159936/formatting-large-currency-or-dollar-values-to-millions-billions
}
